use fmaf for float to avoid conversions

+++ include/replace/backend/ieee.h
@@ -12,7 +12,7 @@ namespace pene
         static void sub_float(float a, float b, float* cptr, void*) noexcept { *cptr = a - b; }
         static void mul_float(float a, float b, float* cptr, void*) noexcept { *cptr = a * b; }
         static void div_float(float a, float b, float* cptr, void*) noexcept { *cptr = a / b; }
-        static void madd_float(float a, float b, float c, float* cptr, void*) noexcept { *cptr=fma(a,b,c); }
+        static void madd_float(float a, float b, float c, float* cptr, void*) noexcept { *cptr=fmaf(a,b,c); }
 
         static void add_double(double a, double b, double* cptr, void*) noexcept { *cptr = a + b; }
         static void sub_double(double a, double b, double* cptr, void*) noexcept { *cptr = a - b; }
